fix signedness comparison

+++ src/database.c
@@ -201,7 +201,7 @@ database_group_get (database_t *database, int64_t id)
   if (!database)
     return VALHALLA_META_GRP_NIL;
 
-  for (i = 0; i < ARRAY_NB_ELEMENTS (g_meta_group); i++)
+  for (i = 0; i < (int) ARRAY_NB_ELEMENTS (g_meta_group); i++)
     if (database->meta_group[i].id == id)
       return i;
 
@@ -228,7 +228,7 @@ database_file_type_get (database_t *database, int64_t id)
   if (!database)
     return VALHALLA_FILE_TYPE_NULL;
 
-  for (i = 0; i < ARRAY_NB_ELEMENTS (g_file_type); i++)
+  for (i = 0; i < (int) ARRAY_NB_ELEMENTS (g_file_type); i++)
     if (database->file_type[i].id == id)
       return i;
 
@@ -296,7 +296,7 @@ database_prepare_stmt (database_t *database)
 
   memcpy (database->stmts, g_stmts, sizeof (g_stmts));
 
-  for (i = 0; i < ARRAY_NB_ELEMENTS (g_stmts); i++)
+  for (i = 0; i < (int) ARRAY_NB_ELEMENTS (g_stmts); i++)
   {
     int res = sqlite3_prepare_v2 (database->db, database->stmts[i].sql,
                                   -1, &database->stmts[i].stmt, NULL);
@@ -973,7 +973,7 @@ vh_database_uninit (database_t *database)
   if (database->path)
     free (database->path);
 
-  for (i = 0; i < ARRAY_NB_ELEMENTS (g_stmts); i++)
+  for (i = 0; i < (int) ARRAY_NB_ELEMENTS (g_stmts); i++)
     if (STMT_GET (i))
       sqlite3_finalize (STMT_GET (i));
 
@@ -1028,10 +1028,10 @@ vh_database_init (const char *path)
   memcpy (database->file_type, g_file_type, sizeof (g_file_type));
   memcpy (database->meta_group, g_meta_group, sizeof (g_meta_group));
 
-  for (i = 0; i < ARRAY_NB_ELEMENTS (g_file_type); i++)
+  for (i = 0; i < (int) ARRAY_NB_ELEMENTS (g_file_type); i++)
     database->file_type[i].id =
       database_type_insert (database, database->file_type[i].name);
-  for (i = 0; i < ARRAY_NB_ELEMENTS (g_meta_group); i++)
+  for (i = 0; i < (int) ARRAY_NB_ELEMENTS (g_meta_group); i++)
     database->meta_group[i].id =
       database_group_insert (database, database->meta_group[i].name);
